job-hunter/requirements.txt
requests
python-dotenv       # secrets
tqdm                # progress bar
APScheduler         # simple cron
slack_sdk           # posting alerts
